Arreglos
FIX: Ahora no se puede agregar evidencia si la actividad tiene la mas alta calificación

Error al traer e Avance cuando no existe calificación (0%)

Se valido la entrada de calificación de evidencia a solo números

Se agregaron campos en BD en programas (Faltan para informe) y ajuste tipo de variable en la Calificación

+++ database/migrations/autoevaluacion/2019_10_28_155956_create_calificaciones_plan_mejoramiento.php
@@ -15,12 +15,12 @@ class CreateCalificacionesPlanMejoramiento extends Migration
     {
         Schema::connection('autoevaluacion')->create('TBL_Califica_Actividades', function (Blueprint $table) {
             $table->increments('PK_CLA_Id');
-            $table->integer('CLA_Calificacion');
+            $table->double('CLA_Calificacion');
             $table->string('CLA_Observacion');
             $table->integer('FK_CLA_Actividad_Mejoramiento')->unsigned();
             $table->integer('FK_CLA_Fecha_Corte')->unsigned();
             $table->timestamps();
-            
+
             $table->foreign("FK_CLA_Actividad_Mejoramiento")->references("PK_ACM_Id")->on("TBL_Actividades_Mejoramiento")->onDelete("cascade");
             $table->foreign("FK_CLA_Fecha_Corte")->references("PK_FCO_Id")->on("TBL_Fechas_corte")->onDelete('cascade');
         });
